Implement a generic state machine and add an example
A generic state machine trait that allows to define any possible
determined finite state machine/reducer, e.g. it allows to provide
state, inputs and outputs of the machine. A wrapper for generic state
machines for practical usage.

An example of the library usage based on the Circuit Breaker pattern.